backend/requirements.txt
# === Web & API ===
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.9
aiofiles
python-dotenv==1.0.0

# === 資料處理 ===
pandas==2.2.3
numpy==1.26.4
scikit-image==0.24.0        # 升級：支援 Python 3.12 + arm64 wheel
Pillow==10.0.0
pyyaml==6.0.3

# === 機器學習核心 ===
torch==2.5.0
torchvision==0.20.0
ultralytics==8.3.40         # 升級：最新穩定版，支援 torch 2.5

# === Hugging Face 生態（必須手動安裝 tokenizers）===
transformers==4.44.2        # 升級：支援 Python 3.12 + tokenizers 0.19.1

# === 電腦視覺進階 ===
git+https://github.com/facebookresearch/segment-anything.git

# === 資料庫 ===
mysql-connector-python
pymysql

# === 安全 ===
passlib==1.7.4
colorama==0.4.6

# === 其他 ===
click==8.1.8
idna==3.10
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
pydantic==2.10.6
pydantic_core==2.27.2
annotated-types==0.7.0
anyio==4.8.0